refactored interface_adapter
