+++ src/interface_adapter/ViewManagerModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.signup;
+package interface_adapter;
 
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
