+++ src/interface_adapter/ViewModel.java
@@ -1,4 +1,4 @@
-package interface_adapter.signup;
+package interface_adapter;
 
 import java.beans.PropertyChangeListener;
