+++ src/interface_adapter/signup/SignupPresenter.java
@@ -1,5 +1,6 @@
 package interface_adapter.signup;
 
+import interface_adapter.ViewManagerModel;
 import use_case.signup.SignupOutputBoundary;
 import use_case.signup.SignupOutputData;
