+++ src/interface_adapter/signup/SignupViewModel.java
@@ -1,5 +1,7 @@
 package interface_adapter.signup;
 
+import interface_adapter.ViewModel;
+
 import java.beans.PropertyChangeListener;
 import java.beans.PropertyChangeSupport;
