+++ src/use_case/login/UserLoginDataAccessInterface.java
@@ -3,7 +3,7 @@ package use_case.login;
 import src.entity.User;
 
 public interface UserLoginDataAccessInterface {
-    boolean existByEmail(String identifier);
+//    boolean existByEmail(String identifier);
     boolean existByName(String identifier);
     void save(User user);
